refactor: split triplet renderer image preprocessing
